+++ custom_model_runner/datarobot_drum/drum/language_predictors/java_predictor/src/main/java/com/datarobot/drum/BasePredictor.java
@@ -1,4 +1,4 @@
-package com.datarobot.custom;
+package com.datarobot.drum;
 
 import java.io.IOException;
