+++ custom_model_runner/datarobot_drum/drum/language_predictors/java_predictor/src/main/java/com/datarobot/drum/ScoringCode.java
@@ -1,4 +1,4 @@
-package com.datarobot.custom;
+package com.datarobot.drum;
 
 import com.datarobot.prediction.IClassificationPredictor;
 import com.datarobot.prediction.IPredictorInfo;
