+++ custom_model_runner/datarobot_drum/drum/language_predictors/java_predictor/src/main/scala/com/datarobot/drum/H2OPredictor.scala
@@ -1,4 +1,4 @@
-package com.datarobot.custom
+package com.datarobot.drum
 
 // h2o dependencies
 import hex.genmodel.GenModel
