+++ custom_model_runner/datarobot_drum/drum/language_predictors/java_predictor/src/main/scala/com/datarobot/drum/H2OPredictorPipeline.scala
@@ -1,4 +1,4 @@
-package com.datarobot.custom
+package com.datarobot.drum
 
 import ai.h2o.mojos.runtime.MojoPipeline;
 import ai.h2o.mojos.runtime.frame.MojoFrame;
